Fix usage of NotImplemented and NotImplementedError.
NotImplemented is a non-callable constant that should be returned from
magic methods used for operator overloading when an operation is
unsupported. This enables Python's type machinery to work properly,
including trying the reflected operation (e.g. using the second
operand's __radd__ instead of the first operand's __add__) and
generating proper TypeError exceptions for unsupported operations.

On the other hand, NotImplementedError is an exception class whose
instances should be raised in abstract methods of abstract base classes
to signal that these methods should be implemented by the child classes.

This commit also raises ValueError and TypeError instead of
NotImplementedError where appropriate (for instance, in the min
function). This mimic's Python's usual behaviour in such cases (for
instance, the built-in min function will raise a TypeError when given
instances of unorderable types in Python 3.x).